feat: add pretty chain name for the "Wrong network..." banner
