+++ shared/wallet/fallback/useErrorMessage.ts
@@ -5,21 +5,23 @@ import { helpers } from 'reef-knot/web3-react';
 import { joinWithOr } from 'utils/join-with-or';
 import { useConnect } from 'wagmi';
 
+import { getPrettyChainName } from 'modules/web3/consts/chains';
+
 export const useErrorMessage = (toActionText?: string): string | undefined => {
   const { isLedger } = useConnectorInfo();
   const {
+    chainId,
     isSupportedChain,
     isChainMatched,
     isAccountActive,
     supportedChainLabels,
-    wagmiChain,
   } = useDappStatus();
   const { error } = useConnect();
 
   // Errors from chain state
 
-  if (isAccountActive && !isChainMatched && wagmiChain) {
-    return `Wrong network. Please switch to ${wagmiChain.name} in your wallet to ${toActionText}.`;
+  if (isAccountActive && !isChainMatched) {
+    return `Wrong network. Please switch to ${getPrettyChainName(chainId)} in your wallet to ${toActionText}.`;
   }
 
   if (!isSupportedChain) {
