+++ utils/etherscan.ts
@@ -1,5 +1,5 @@
 import invariant from 'tiny-invariant';
-import { wagmiChainMap } from 'modules/web3/web3-provider/web3-provider';
+import { wagmiChainMap } from 'modules/web3/consts/chains';
 
 export enum ETHERSCAN_ENTITIES {
   TX = 'tx',
